changes to prompting
Added option to include messages array, but removed easy access to systemPrompt. This can now be accessed by providing a messages array. The imageToText function is not yet updated in this style, as this needs more thought --> the processed image file needs to be in this array. Also this commit removed the messageHistory entirely to make functions more clean. Chat history functionality must now be handled by the user, which is possible by providing the messages array.